% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_state.R
\name{check_state}
\alias{check_state}
\title{Checks to see if user input a valid state.}
\usage{
check_state(state)
}
\arguments{
\item{state}{A character input that is either a state name (e.g., "North Carolina" or "Florida)
or a postal code ("FL" or "nc"). Can be capitalized or not.}
}
\value{
An error if the was inputted incorrectly.
}
\description{
Checks to see if user input a valid state.
}
