% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_formatter.R
\name{col_formatter}
\alias{col_formatter}
\title{Function that takes in information from the yaml to format each column appropriately.}
\usage{
col_formatter(cols, column_names, date_format)
}
\arguments{
\item{cols}{The column classes as ascertained by load_state().}

\item{column_names}{The column names, as ascertained by load_state()}

\item{date_format}{The date format for date-type files, as ascertained by load_state()}
}
\value{
R script to be evaluated by load_state() that will read the columns as the correct classes.
}
\description{
Function that takes in information from the yaml to format each column appropriately.
}
