% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_file_case.R
\name{multiple_file_case}
\alias{multiple_file_case}
\title{A function to read state voter record datasets split across several different files.}
\usage{
multiple_file_case(
  path,
  file_type,
  yaml,
  col_classes,
  col_names,
  d_format,
  delim
)
}
\description{
A function to read state voter record datasets split across several different files.
}
