% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_yaml.R
\name{retrieve_yaml}
\alias{retrieve_yaml}
\title{Function to take in a state and fetch its associated yaml.}
\usage{
retrieve_yaml(state, history = FALSE)
}
\arguments{
\item{state}{A state name or postal code (e.g., "Florida" or "FL" or "fL")}
}
\value{
The YAML file on file for this state or an error, alerting users that it is not currently documented.
}
\description{
Function to take in a state and fetch its associated yaml.
}
