% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_file_case.R
\name{single_file_case}
\alias{single_file_case}
\title{Helper function for load_state() dedicated to states with a single file.
Uses col_formatter() to determine the formats of each column in the dataset and
then reads-in a single dataframe based off of those specifications--as well as others
in the embedded YAML data.}
\usage{
single_file_case(col_classes, col_names, d_format, yaml, vtr_file)
}
\arguments{
\item{col_classes}{Column classes, as ascertained from the YAML by load_state()}

\item{col_names}{Column names, as ascertained from the YAML by load_state()}

\item{d_format}{The format of the date-fields, as ascertained from the YAML by load_state()}
}
\value{
A data frame of the state's voter registration file.
}
\description{
Helper function for load_state() dedicated to states with a single file.
Uses col_formatter() to determine the formats of each column in the dataset and
then reads-in a single dataframe based off of those specifications--as well as others
in the embedded YAML data.
}
